docs: avoid converts's use of uniconvertor
ImageMagick's convert uses uniconvertor to process figs, which is a
python 2 project with unclear future.

Instead, use fig2dev which is used in other places already, although
differently for the two affected figs:
- pagesetup2.fig is converted to pdf by the Makefile already, using
  fig2dev
- processus.eps is a conversion of processus.fig using fig2dev; both are
  shipped in-tree

We leave this different treatment as is but avoid convert from the pdf
to png instead of from the fig now.