2-3: except Exception as Var
python2 -3 warns about python2 constructs which are deprecated for
python2 because they are incompatible with python3. Replace them to get
as close as possible to a python2+python3 codebase.

First step: 2to3 -f except
Turn "except Exception, Var" into "except Exception as Var".

Note that we treat the entire codebase simultaneously in these steps,
even though parts such as lib/contrib/which will not be used with
python3.