2-3: relative imports
Adjust relative imports so that under python 2 they are declared to be
relative, and under python 3 they are still valid.

We prefer the short form ".module" over the full "package.subdir.module"
because it is closer to the original. Grubber modules need the full form
because of the way they are called.